Make Grid object mutable.
As it doesn't seem that a fully immutable object is possible in general (even
if it kinda works in this case, since there's nothing that needs to be hidden),
since all relevant data must pass through a public constructor.

Is a private constructor possible? If so, immutable objects can be doable.